pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-video-editor"
version = "0.1.0"
description = "AI-powered video editing tool using Ollama and FFMPEG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ollama>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typer>=0.15.0",
    "rich>=13.0.0",
    "static-ffmpeg>=2.5",
    "send2trash>=1.8.0",
]

[project.optional-dependencies]
audio = ["openai-whisper>=20231117"]
all = ["openai-whisper>=20231117"]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
ai-video-editor = "ai_video_editor.__main__:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"